Move database init from server.ts to index.ts
